PCA - ExtensionExecutor plugin improvments.
