+++ components/org.apache.stratos.python.cartridge.agent/src/main/python/cartridge.agent/cartridge.agent/constants.py
@@ -125,7 +125,7 @@ DOMAIN_MAPPING_REMOVED_EVENT = "DomainMappingRemovedEvent"
 MEMBER_INITIALIZED_EVENT = "MemberInitializedEvent"
 MEMBER_ACTIVATED_EVENT = "MemberActivatedEvent"
 MEMBER_TERMINATED_EVENT = "MemberTerminatedEvent"
-MEMBER_SUSPENDED_EVENT = "MembeSuspendedEvent"
+MEMBER_SUSPENDED_EVENT = "MemberSuspendedEvent"
 MEMBER_STARTED_EVENT = "MemberStartedEvent"
 TENANT_SUBSCRIBED_EVENT = "TenantSubscribedEvent"
 APPLICATION_SIGNUP_REMOVAL_EVENT = "ApplicationSignUpRemovedEvent"
